Refaktor litt

+++ src/main/kotlin/no/nav/faktureringskomponenten/domain/models/FakturaMottatt.kt
@@ -13,7 +13,7 @@ class FakturaMottatt(
     val id: Long? = null,
 
     @Column(name = "faktura_referanse_nr", nullable = false)
-    val fakturaReferanseNr: Int? = null,
+    val fakturaReferanseNr: Long? = null,
 
     @Column(name = "faktura_nummer", nullable = true)
     val fakturaNummer: String? = null,
